Componentize key ceremony listing and show in admin home
