+++ src/electionguard_gui/web/components/admin/admin-home-component.js
@@ -1,4 +1,9 @@
+import KeyCeremonyList from "../shared/key-ceremony-list-component.js";
+
 export default {
+  components: {
+    KeyCeremonyList,
+  },
   template: /*html*/ `
   <div class="container col-6">
     <div class="text-center mb-4">
@@ -12,5 +17,11 @@ export default {
         <a href="#/admin/setup-election" class="btn btn-primary">Setup Election</a>
       </div>
     </div>
-  </div>`,
+  </div>
+  <div class="text-center mt-4">
+    <h2>Key Ceremonies</h2>
+    <key-ceremony-list :is-admin="true"></key-ceremony-list>
+  </div>
+
+  `,
 };
