+++ src/electionguard_gui/web/components/shared/key-ceremony-list-component.js
@@ -1,0 +1,41 @@
+import RouterService from "/services/router-service.js";
+
+export default {
+  props: {
+    isAdmin: Boolean,
+  },
+  data() {
+    return {
+      keyCeremonies: [],
+    };
+  },
+  methods: {
+    key_ceremonies_found: function (key_ceremonies) {
+      this.keyCeremonies = key_ceremonies;
+    },
+    getKeyCeremonyUrl: function (keyCeremony) {
+      const page = this.isAdmin
+        ? RouterService.routes.viewKeyCeremonyAdminPage
+        : RouterService.routes.viewKeyCeremonyGuardianPage;
+      return RouterService.getUrl(page, {
+        keyCeremonyId: keyCeremony.id,
+      });
+    },
+  },
+  async mounted() {
+    console.log("begin watching for key ceremonies");
+    eel.expose(this.key_ceremonies_found, "key_ceremonies_found");
+    eel.watch_key_ceremonies();
+  },
+  unmounted() {
+    console.log("stop watching key ceremonies");
+    eel.stop_watching_key_ceremonies();
+  },
+  template: /*html*/ `
+  <p v-if="!keyCeremonies">No key ceremonies found...</p>
+
+  <div v-if="keyCeremonies" class="d-grid gap-2 d-md-block">
+    <a :href="getKeyCeremonyUrl(keyCeremony)" v-for="keyCeremony in keyCeremonies" class="btn btn-primary me-2 mt-2">{{ keyCeremony.key_ceremony_name }}</a>
+  </div>
+  `,
+};
